split Result and Attack again, importing everything early in __init__ fixes the circular dep issue.
